feat: enhance Transformer model with dynamic GPU memory allocation and improved data type handling

+++ supervised_learning/transformer_apps/5-transformer.py
@@ -152,7 +152,9 @@ class Encoder(tf.keras.layers.Layer):
         self.N = N
         self.dm = dm
         self.embedding = tf.keras.layers.Embedding(input_vocab, dm)
-        self.positional_encoding = positional_encoding(max_seq_len, dm)
+        self.positional_encoding = tf.cast(
+            positional_encoding(max_seq_len, dm), dtype=tf.float32
+        )
         self.blocks = [
             EncoderBlock(dm, h, hidden, drop_rate) for _ in range(N)
         ]
@@ -163,7 +165,7 @@ class Encoder(tf.keras.layers.Layer):
         seq_len = tf.shape(x)[1]
         x = self.embedding(x)
         x *= tf.math.sqrt(tf.cast(self.dm, tf.float32))
-        x += self.positional_encoding[:seq_len, :]
+        x += tf.cast(self.positional_encoding[:seq_len, :], dtype=x.dtype)
         x = self.dropout(x, training=training)
         for i in range(self.N):
             x = self.blocks[i](x, training, mask)
@@ -181,7 +183,9 @@ class Decoder(tf.keras.layers.Layer):
         self.N = N
         self.dm = dm
         self.embedding = tf.keras.layers.Embedding(target_vocab, dm)
-        self.positional_encoding = positional_encoding(max_seq_len, dm)
+        self.positional_encoding = tf.cast(
+            positional_encoding(max_seq_len, dm), dtype=tf.float32
+        )
         self.blocks = [
             DecoderBlock(dm, h, hidden, drop_rate) for _ in range(N)
         ]
@@ -192,7 +196,7 @@ class Decoder(tf.keras.layers.Layer):
         seq_len = tf.shape(x)[1]
         x = self.embedding(x)
         x *= tf.math.sqrt(tf.cast(self.dm, tf.float32))
-        x += self.positional_encoding[:seq_len, :]
+        x += tf.cast(self.positional_encoding[:seq_len, :], dtype=x.dtype)
         x = self.dropout(x, training=training)
         for i in range(self.N):
             x = self.blocks[i](
